lib: cockpit: rename location to location-utils
Prepare for moving cockpit.location to a separate typescript file named
location.
