+++ pkg/lib/cockpit/_internal/location-utils.ts
